Standard preprocessing
