+++ rllm/preprocessing/__init__.py
@@ -1,15 +1,4 @@
-from rllm.preprocessing._fillna import (
-    fillna_numerical,
-    fillna_categorical,
-    fillna_binary,
-    fillna_by_coltype,
-)
-from rllm.preprocessing._type_convert import (
-    encode_categorical,
-    convert_binary,
-    dict_to_df,
-)
-from rllm.preprocessing._text_tokenize import (
+from rllm.preprocessing.text_tokenize import (
     TokenizerConfig,
     process_tokenized_column,
     tokenize_strings,
@@ -17,25 +6,17 @@ from rllm.preprocessing._text_tokenize import (
     tokenize_merged_cols,
     save_column_name_tokens,
 )
-from rllm.preprocessing._word_embedding import (
+from rllm.preprocessing.word_embedding import (
     TextEmbedderConfig,
     embed_text_column,
 )
-from rllm.preprocessing._timestamp import TimestampPreprocessor
+from rllm.preprocessing.data_clean import preprocess_numerical_string
+from rllm.preprocessing.timestamp import TimestampPreprocessor
 from rllm.preprocessing.df_to_tensor import df_to_tensor
 
 __all__ = [
     # df to tensor
     "df_to_tensor",
-    # fillna
-    "fillna_numerical",
-    "fillna_categorical",
-    "fillna_binary",
-    "fillna_by_coltype",
-    # type convert
-    "encode_categorical",
-    "convert_binary",
-    "dict_to_df",
     # text tokenize
     "TokenizerConfig",
     "process_tokenized_column",
@@ -48,4 +29,6 @@ __all__ = [
     "embed_text_column",
     # timestamp
     "TimestampPreprocessor",
+    # data clean
+    "preprocess_numerical_string",
 ]
