+++ rllm/preprocessing/df_to_tensor.py
@@ -1,6 +1,5 @@
 from typing import Optional, Union, Tuple
 from pandas import Series
-import re
 
 import torch
 from torch import Tensor
@@ -10,73 +9,18 @@ from rllm.preprocessing._type_convert import (
     encode_categorical,
     convert_binary,
 )
-from rllm.preprocessing._text_tokenize import (
+from rllm.preprocessing.data_clean import preprocess_numerical_string
+from rllm.preprocessing.text_tokenize import (
     TokenizerConfig,
     process_tokenized_column,
     tokenize_merged_cols,
 )
-from rllm.preprocessing._word_embedding import (
+from rllm.preprocessing.word_embedding import (
     TextEmbedderConfig,
     embed_text_column,
 )
-from rllm.preprocessing._timestamp import TimestampPreprocessor
+from rllm.preprocessing.timestamp import TimestampPreprocessor
 from rllm.types import ColType
-
-
-def _clean_numerical_value(val):
-    """
-    Clean a single numerical value that may contain special formatting.
-
-    Handles:
-    - Currency symbols: $132.42, ¥1000, €50.5, etc.
-    - Percentages: 1.572%, 25% (converts to decimal by dividing by 100)
-    - Thousand separators: 1,234.56
-    - Scientific notation: 1.5e10, 2.3E-5
-    - Whitespace: "  123.45  ", "1 234.56"
-
-    Args:
-        val: Value to clean (can be str, int, float, or None)
-
-    Returns:
-        float or None: Cleaned numerical value or None if invalid
-    """
-    # Handle None and existing numeric types
-    if val is None or isinstance(val, (int, float)):
-        return val
-
-    # Convert to string and strip whitespace
-    val_str = str(val).strip()
-    if not val_str:
-        return None
-
-    try:
-        # Check for percentage format
-        is_percentage = '%' in val_str
-        if is_percentage:
-            val_str = val_str.replace('%', '')
-        # Remove all whitespace (handles "1 234.56" format)
-        val_str = val_str.replace(' ', '')
-        # Remove currency symbols and other non-numeric chars
-        # Keep: digits, decimal point, minus sign, comma, and 'e/E' for scientific notation
-        val_str = re.sub(r'[^\d.\-,eE]', '', val_str)
-        # Remove thousand separators
-        val_str = val_str.replace(',', '')
-        # Validate and convert to float
-        if val_str and val_str not in ['-', '.', '-.', 'e', 'E']:
-            num_val = float(val_str)
-            # Convert percentage to decimal
-            if is_percentage:
-                num_val = num_val / 100.0
-            return num_val
-        return None
-
-    except (ValueError, AttributeError):
-        # Return None for invalid values (will be handled by fillna)
-        return None
-
-
-def preprocess_numerical_string(col_series: Series) -> Series:
-    return col_series.map(_clean_numerical_value)
 
 
 def df_to_tensor(
@@ -95,11 +39,19 @@ def df_to_tensor(
         col_types: Dictionary mapping column names to column types
         target_col: Name of target column
         text_embedder_config: Configuration for text embedding
-        tokenizer_config: Configuration for tokenization
+        tokenizer_config: Configuration for tokenization; if provided and
+            ``tokenize_combine`` is True, all TEXT columns are jointly tokenized
+            and stored as a single entry in ``feat_dict[ColType.TEXT]``.
+        concat: Whether to concatenate/stack features of the same column type
+            (e.g., numerical and categorical along the last dim, text and
+            timestamp along the feature/channel dim).
+        cat_hardcode: Whether to cast categorical features to integer type.
 
     Returns:
         tuple: (feat_dict, y) where feat_dict contains feature tensors by column type,
-            and y is the target tensor (None if no target_col)
+            and y is the target tensor (None if no target_col). When TEXT columns
+            are tokenized, the corresponding value is a tuple of
+            ``(input_ids, attention_mask)``; otherwise it is an embedded tensor.
     """
 
     # 1. Iterate each column
